# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SalesERP API",
    "description": "Sales ERP Management System API",
    "version": "1.0.0"
  },
  "host": "erp-sales-backend.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/sales/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "paymentStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customerName": {
                  "example": "any"
                },
                "customerContact": {
                  "example": "any"
                },
                "invoiceNo": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/sales/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customerName": {
                  "example": "any"
                },
                "customerContact": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/sales/{id}/payment": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "amountPaid": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/sales/soft-delete/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/sales/restore/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/sales/deleted/list": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}